+++ src/web/src/components/index.ts
@@ -14,8 +14,8 @@ export { AnimationStyles } from "./atoms/AnimationStyles";
 export { SynthBackground } from "./atoms/SynthBackground";
 export { Skeleton, CardSkeleton, RowSkeleton } from "./atoms/Skeleton";
 export { ActivityIndicator } from "./atoms/ActivityIndicator";
-export { ProgressBar } from "./atoms/ProgressBar";
-export type { ProgressBarSegment, ProgressBarProps } from "./atoms/ProgressBar";
+export { ProgressBar } from "@4lt7ab/ui/ui";
+export type { ProgressBarSegment, ProgressBarProps } from "@4lt7ab/ui/ui";
 export { ReferenceTypeBadge } from "./atoms/ReferenceTypeBadge";
 // Molecules
 export { Card } from "./molecules/Card";